remove 2 variable shadow possible errors switch from i64 to i32 but still not correct (for us OK as the numbers will always be small)

+++ src/types.rs
@@ -59,8 +59,8 @@ pub fn generate_random_vec_u8(size: usize) -> Vec<u8> {
 }
 
 /// Quorum factor.
-pub const QUORUM_FACTOR: f64 = ::kademlia_routing_table::QUORUM_SIZE as f64 /
-                               ::kademlia_routing_table::GROUP_SIZE as f64;
+pub const QUORUM_FACTOR: f32 = ::kademlia_routing_table::QUORUM_SIZE as f32 /
+                               ::kademlia_routing_table::GROUP_SIZE as f32;
 
 #[derive(PartialEq, Eq, PartialOrd, Ord, Hash, Clone, RustcEncodable, RustcDecodable)]
 /// Address.
